third_party/WebKit/LayoutTests/fast/html/clone-keygen-expected.txt

CONSOLE WARNING: The <keygen> element is deprecated and will be removed in M57, around March 2017. See https://www.chromestatus.com/features/5716060992962560 for more details.
Tests that a cloned keygen has the right shadow pseudo ID. You should see two big keygens with purple text (they should look the same.)


PASS specialK.clientHeight is k.clientHeight
PASS internals.shadowPseudoId(cloneShadow.firstChild) is internals.shadowPseudoId(kShadow.firstChild)
PASS successfullyParsed is true

TEST COMPLETE
